+++ root/etc/cont-init.d/50-setup.sh
@@ -7,7 +7,7 @@
 # Setup backup
 if [ ! -e "/config/backup" ]; then
 	mkdir /config/backup
-  chown abc:abc /config/backup
+  [ -n "$NO_CHOWN" ] || chown abc:abc /config/backup
 fi
 
 #
@@ -17,7 +17,7 @@ fi
 if [ ! -d "/config/openvpn" ]; then
   echo "Creating /config/openvpn"
   mkdir -p /config/openvpn
-  chown abc:abc /config/openvpn
+  [ -n "$NO_CHOWN" ] || chown abc:abc /config/openvpn
 fi
 
 # Check directories inside openvpn directory
@@ -27,7 +27,7 @@ do
 	if [ ! -d "/config/openvpn/$h" ]; then
 		echo "Creating /config/openvpn/$h"
 		mkdir /config/openvpn/$h
-    chown abc:abc /config/openvpn/$h
+    [ -n "$NO_CHOWN" ] || chown abc:abc /config/openvpn/$h
 	fi
 done
 
@@ -49,7 +49,7 @@ for h in "${HOOKS_DIR[@]}"; do
 	if [ ! -d "/config/openvpn/hooks/$h" ]; then
 		echo "Creating /config/openvpn/hooks/$h"
 		mkdir /config/openvpn/hooks/$h
-    chown abc:abc /config/openvpn/hooks/$h
+    [ -n "$NO_CHOWN" ] || chown abc:abc /config/openvpn/hooks/$h
 	fi
 done
 
@@ -67,17 +67,17 @@ fi
 if [ ! -d "/config/ssl" ]; then
 	echo "Setting up /config/ssl"
 	mkdir -p /config/ssl
-  chown abc:abc /config/ssl
+  [ -n "$NO_CHOWN" ] || chown abc:abc /config/ssl
 fi
 
 if [ ! -e "$EASYRSA_VARS_FILE" ]; then
   #cp -R -u $EASYRSA/openssl-easyrsa.cnf $EASYRSA_SSL_CONF
 	cp -R -u $EASYRSA/vars.example $EASYRSA_VARS_FILE
-  chown abc:abc $EASYRSA_VARS_FILE
+  [ -n "$NO_CHOWN" ] || chown abc:abc $EASYRSA_VARS_FILE
 fi
 
 # Setup tmp
 if [ ! -e "/config/tmp" ]; then
 	mkdir /config/tmp
-  chown abc:abc /config/tmp
+  [ -n "$NO_CHOWN" ] || chown abc:abc /config/tmp
 fi