+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Moved `tmp-dir` to server related config.
 - Fixed bug in dockumentation (GUID -> PGID).
 - Added `SKIP_APP` parameter (to skip app startup).
+- Added `NO_CHOWN` parameter (don't fix config files permissions).
 
 ### 2.0.3 - Improved IPv6 support & finalized client mode support